+++ vscode/src/main.ts
@@ -107,7 +107,7 @@ const register = async (
         disposables.push(vscode.workspace.onDidChangeTextDocument(updateParseTreeOnEdit))
     }
 
-    const symfRunner = platform.createSymfRunner?.(config.experimentalSymfPath, config.experimentalSymfAnthropicKey)
+    const symfRunner = platform.createSymfRunner?.(context, config.experimentalSymfAnthropicKey)
 
     const {
         featureFlagProvider,
@@ -355,7 +355,11 @@ const register = async (
         // Register URI Handler (vscode://sourcegraph.cody-ai)
         vscode.window.registerUriHandler({
             handleUri: async (uri: vscode.Uri) => {
-                await authProvider.tokenCallbackHandler(uri, config.customHeaders)
+                if (uri.path === '/app-done') {
+                    await sidebarChatProvider.simplifiedOnboardingReloadEmbeddingsState()
+                } else {
+                    await authProvider.tokenCallbackHandler(uri, config.customHeaders)
+                }
             },
         }),
         statusBar,