+++ vscode/webviews/Notices/EmbeddingsEnabledNotice.story.tsx
@@ -1,0 +1,23 @@
+import { Meta, StoryObj } from '@storybook/react'
+
+import { VSCodeStoryDecorator } from '../storybook/VSCodeStoryDecorator'
+
+import { EmbeddingsEnabledNotice } from './EmbeddingsEnabledNotice'
+
+import styles from '../storybook/VSCodeStoryDecorator.module.css'
+
+const meta: Meta<typeof EmbeddingsEnabledNotice> = {
+    title: 'cody/App-less Onboarding',
+    component: EmbeddingsEnabledNotice,
+    decorators: [VSCodeStoryDecorator],
+}
+
+export default meta
+
+export const EmbeddingsEnabled: StoryObj<typeof EmbeddingsEnabledNotice> = {
+    render: () => (
+        <div className={styles.testDarkSidebar}>
+            <EmbeddingsEnabledNotice />
+        </div>
+    ),
+}